requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
opencv-python>=4.9.0
ultralytics>=8.2.0
numpy>=1.23.0
gspread>=5.12.0
google-auth>=2.27.0
google-api-python-client>=2.110.0
